DRTVWR-418: Tweak 'darwin' platform in SConstruct to honor CCFLAGS.
We carefully set the CCFLAGS environment variable before invoking scons, but
the SConstruct file unconditionally stomps on it. Makes you appreciate the
funny 'make' semantics in which variable settings in the Makefile provide
defaults that can be overridden by command-line switches or environment
variables.

Change SConstruct to (at least in the 'darwin' case) set its CCFLAGS to the
CCFLAGS environment variable we explicitly set in build-cmd.sh.